Started writing the networked piga input system.
This system uses enet as network library to transmit inputs from a
variable number of providing clients. The client uses the same
SharedLibWrapper as the host does and loads hosts in the same way too.

For packet construction and versioning, google protocol buffers is
used. There will eventually be more networked APIs which use the same
interface as the inputs.